refine pre-commit

+++ ppdiffusers/ppdiffusers/models/t5_film_transformer.py
@@ -26,31 +26,30 @@ from .modeling_utils import ModelMixin
 class T5FilmDecoder(ModelMixin, ConfigMixin):
     @register_to_config
     def __init__(
-            self,
-            input_dims: int=128,
-            targets_length: int=256,
-            max_decoder_noise_time: float=2000.0,
-            d_model: int=768,
-            num_layers: int=12,
-            num_heads: int=12,
-            d_kv: int=64,
-            d_ff: int=2048,
-            dropout_rate: float=0.1, ):
+        self,
+        input_dims: int = 128,
+        targets_length: int = 256,
+        max_decoder_noise_time: float = 2000.0,
+        d_model: int = 768,
+        num_layers: int = 12,
+        num_heads: int = 12,
+        d_kv: int = 64,
+        d_ff: int = 2048,
+        dropout_rate: float = 0.1,
+    ):
         super().__init__()
 
         self.conditioning_emb = nn.Sequential(
-            nn.Linear(
-                d_model, d_model * 4, bias_attr=False),
+            nn.Linear(d_model, d_model * 4, bias_attr=False),
             nn.Silu(),
-            nn.Linear(
-                d_model * 4, d_model * 4, bias_attr=False),
-            nn.Silu(), )
+            nn.Linear(d_model * 4, d_model * 4, bias_attr=False),
+            nn.Silu(),
+        )
 
         self.position_encoding = nn.Embedding(targets_length, d_model)
         self.position_encoding.weight.stop_gradient = True
 
-        self.continuous_inputs_projection = nn.Linear(
-            input_dims, d_model, bias_attr=False)
+        self.continuous_inputs_projection = nn.Linear(input_dims, d_model, bias_attr=False)
 
         self.dropout = nn.Dropout(p=dropout_rate)
 
@@ -62,7 +61,8 @@ class T5FilmDecoder(ModelMixin, ConfigMixin):
                 d_kv=d_kv,
                 num_heads=num_heads,
                 d_ff=d_ff,
-                dropout_rate=dropout_rate, )
+                dropout_rate=dropout_rate,
+            )
             self.decoders.append(lyr)
 
         self.decoder_norm = T5LayerNorm(d_model)
@@ -71,13 +71,10 @@ class T5FilmDecoder(ModelMixin, ConfigMixin):
         self.spec_out = nn.Linear(d_model, input_dims, bias_attr=False)
 
     def encoder_decoder_mask(self, query_input, key_input):
-        mask = paddle.multiply(
-            query_input.unsqueeze(-1),
-            key_input.unsqueeze(-2).cast(query_input.dtype))
+        mask = paddle.multiply(query_input.unsqueeze(-1), key_input.unsqueeze(-2).cast(query_input.dtype))
         return mask.unsqueeze(-3)
 
-    def forward(self, encodings_and_masks, decoder_input_tokens,
-                decoder_noise_time):
+    def forward(self, encodings_and_masks, decoder_input_tokens, decoder_noise_time):
         batch, _, _ = decoder_input_tokens.shape
         assert decoder_noise_time.shape[0] == batch
 
@@ -85,7 +82,8 @@ class T5FilmDecoder(ModelMixin, ConfigMixin):
         time_steps = get_timestep_embedding(
             decoder_noise_time * self.config.max_decoder_noise_time,
             embedding_dim=self.config.d_model,
-            max_period=self.config.max_decoder_noise_time, ).cast(self.dtype)
+            max_period=self.config.max_decoder_noise_time,
+        ).cast(self.dtype)
 
         conditioning_emb = self.conditioning_emb(time_steps).unsqueeze(1)
 
@@ -96,37 +94,34 @@ class T5FilmDecoder(ModelMixin, ConfigMixin):
         # If we want to use relative positions for audio context, we can just offset
         # this sequence by the length of encodings_and_masks.
         decoder_positions = paddle.broadcast_to(
-            paddle.arange(seq_length, ),
-            shape=(batch, seq_length), )
+            paddle.arange(
+                seq_length,
+            ),
+            shape=(batch, seq_length),
+        )
 
         position_encodings = self.position_encoding(decoder_positions)
-        inputs = self.continuous_inputs_projection(
-            decoder_input_tokens.cast(position_encodings.dtype))
+        inputs = self.continuous_inputs_projection(decoder_input_tokens.cast(position_encodings.dtype))
         inputs += position_encodings
         y = self.dropout(inputs)
 
         # decoder: No padding present.
-        decoder_mask = paddle.ones(
-            decoder_input_tokens.shape[:2], dtype=inputs.dtype)
+        decoder_mask = paddle.ones(decoder_input_tokens.shape[:2], dtype=inputs.dtype)
 
         # Translate encoding masks to encoder-decoder masks.
-        encodings_and_encdec_masks = [
-            (x, self.encoder_decoder_mask(decoder_mask, y))
-            for x, y in encodings_and_masks
-        ]
+        encodings_and_encdec_masks = [(x, self.encoder_decoder_mask(decoder_mask, y)) for x, y in encodings_and_masks]
 
         # cross attend style: concat encodings
-        encoded = paddle.concat(
-            [x[0] for x in encodings_and_encdec_masks], axis=1)
-        encoder_decoder_mask = paddle.concat(
-            [x[1] for x in encodings_and_encdec_masks], axis=-1)
+        encoded = paddle.concat([x[0] for x in encodings_and_encdec_masks], axis=1)
+        encoder_decoder_mask = paddle.concat([x[1] for x in encodings_and_encdec_masks], axis=-1)
 
         for lyr in self.decoders:
             y = lyr(
                 y,
                 conditioning_emb=conditioning_emb,
                 encoder_hidden_states=encoded,
-                encoder_attention_mask=encoder_decoder_mask, )[0]
+                encoder_attention_mask=encoder_decoder_mask,
+            )[0]
 
         y = self.decoder_norm(y)
         y = self.post_dropout(y)
@@ -136,13 +131,7 @@ class T5FilmDecoder(ModelMixin, ConfigMixin):
 
 
 class DecoderLayer(nn.Layer):
-    def __init__(self,
-                 d_model,
-                 d_kv,
-                 num_heads,
-                 d_ff,
-                 dropout_rate,
-                 layer_norm_epsilon=1e-6):
+    def __init__(self, d_model, d_kv, num_heads, d_ff, dropout_rate, layer_norm_epsilon=1e-6):
         super().__init__()
         self.layer = nn.LayerList()
 
@@ -152,7 +141,9 @@ class DecoderLayer(nn.Layer):
                 d_model=d_model,
                 d_kv=d_kv,
                 num_heads=num_heads,
-                dropout_rate=dropout_rate, ))
+                dropout_rate=dropout_rate,
+            )
+        )
 
         # cross attention: layer 1
         self.layer.append(
@@ -161,7 +152,9 @@ class DecoderLayer(nn.Layer):
                 d_kv=d_kv,
                 num_heads=num_heads,
                 dropout_rate=dropout_rate,
-                layer_norm_epsilon=layer_norm_epsilon, ))
+                layer_norm_epsilon=layer_norm_epsilon,
+            )
+        )
 
         # Film Cond MLP + dropout: last layer
         self.layer.append(
@@ -169,62 +162,67 @@ class DecoderLayer(nn.Layer):
                 d_model=d_model,
                 d_ff=d_ff,
                 dropout_rate=dropout_rate,
-                layer_norm_epsilon=layer_norm_epsilon, ))
+                layer_norm_epsilon=layer_norm_epsilon,
+            )
+        )
 
     def forward(
-            self,
-            hidden_states,
-            conditioning_emb=None,
-            attention_mask=None,
-            encoder_hidden_states=None,
-            encoder_attention_mask=None,
-            encoder_decoder_position_bias=None, ):
+        self,
+        hidden_states,
+        conditioning_emb=None,
+        attention_mask=None,
+        encoder_hidden_states=None,
+        encoder_attention_mask=None,
+        encoder_decoder_position_bias=None,
+    ):
         hidden_states = self.layer[0](
             hidden_states,
             conditioning_emb=conditioning_emb,
-            attention_mask=attention_mask, )
+            attention_mask=attention_mask,
+        )
 
         if encoder_hidden_states is not None:
-            encoder_extended_attention_mask = paddle.where(
-                encoder_attention_mask > 0, 0.0,
-                -1e10).cast(encoder_hidden_states.dtype)
+            encoder_extended_attention_mask = paddle.where(encoder_attention_mask > 0, 0.0, -1e10).cast(
+                encoder_hidden_states.dtype
+            )
 
             hidden_states = self.layer[1](
                 hidden_states,
                 key_value_states=encoder_hidden_states,
-                attention_mask=encoder_extended_attention_mask, )
+                attention_mask=encoder_extended_attention_mask,
+            )
 
         # Apply Film Conditional Feed Forward layer
         hidden_states = self.layer[-1](hidden_states, conditioning_emb)
 
-        return (hidden_states, )
+        return (hidden_states,)
 
 
 class T5LayerSelfAttentionCond(nn.Layer):
     def __init__(self, d_model, d_kv, num_heads, dropout_rate):
         super().__init__()
         self.layer_norm = T5LayerNorm(d_model)
-        self.FiLMLayer = T5FiLMLayer(
-            in_features=d_model * 4, out_features=d_model)
+        self.FiLMLayer = T5FiLMLayer(in_features=d_model * 4, out_features=d_model)
         self.attention = Attention(
             query_dim=d_model,
             heads=num_heads,
             dim_head=d_kv,
             out_bias=False,
-            scale_qk=False, )
+            scale_qk=False,
+        )
         self.dropout = nn.Dropout(dropout_rate)
 
     def forward(
-            self,
-            hidden_states,
-            conditioning_emb=None,
-            attention_mask=None, ):
+        self,
+        hidden_states,
+        conditioning_emb=None,
+        attention_mask=None,
+    ):
         # pre_self_attention_layer_norm
         normed_hidden_states = self.layer_norm(hidden_states)
 
         if conditioning_emb is not None:
-            normed_hidden_states = self.FiLMLayer(normed_hidden_states,
-                                                  conditioning_emb)
+            normed_hidden_states = self.FiLMLayer(normed_hidden_states, conditioning_emb)
 
         # Self-attention block
         attention_output = self.attention(normed_hidden_states)
@@ -235,28 +233,30 @@ class T5LayerSelfAttentionCond(nn.Layer):
 
 
 class T5LayerCrossAttention(nn.Layer):
-    def __init__(self, d_model, d_kv, num_heads, dropout_rate,
-                 layer_norm_epsilon):
+    def __init__(self, d_model, d_kv, num_heads, dropout_rate, layer_norm_epsilon):
         super().__init__()
         self.attention = Attention(
             query_dim=d_model,
             heads=num_heads,
             dim_head=d_kv,
             out_bias=False,
-            scale_qk=False, )
+            scale_qk=False,
+        )
         self.layer_norm = T5LayerNorm(d_model, eps=layer_norm_epsilon)
         self.dropout = nn.Dropout(dropout_rate)
 
     def forward(
-            self,
-            hidden_states,
-            key_value_states=None,
-            attention_mask=None, ):
+        self,
+        hidden_states,
+        key_value_states=None,
+        attention_mask=None,
+    ):
         normed_hidden_states = self.layer_norm(hidden_states)
         attention_output = self.attention(
             normed_hidden_states,
             encoder_hidden_states=key_value_states,
-            attention_mask=attention_mask.squeeze(1), )
+            attention_mask=attention_mask.squeeze(1),
+        )
         layer_output = hidden_states + self.dropout(attention_output)
         return layer_output
 
@@ -264,8 +264,7 @@ class T5LayerCrossAttention(nn.Layer):
 class T5LayerFFCond(nn.Layer):
     def __init__(self, d_model, d_ff, dropout_rate, layer_norm_epsilon):
         super().__init__()
-        self.DenseReluDense = T5DenseGatedActDense(
-            d_model=d_model, d_ff=d_ff, dropout_rate=dropout_rate)
+        self.DenseReluDense = T5DenseGatedActDense(d_model=d_model, d_ff=d_ff, dropout_rate=dropout_rate)
         self.film = T5FiLMLayer(in_features=d_model * 4, out_features=d_model)
         self.layer_norm = T5LayerNorm(d_model, eps=layer_norm_epsilon)
         self.dropout = nn.Dropout(dropout_rate)
@@ -306,9 +305,7 @@ class T5LayerNorm(nn.Layer):
 
     def __init__(self, hidden_size, eps=1e-6):
         super().__init__()
-        self.weight = self.create_parameter(
-            shape=[hidden_size],
-            default_initializer=nn.initializer.Constant(1.0))
+        self.weight = self.create_parameter(shape=[hidden_size], default_initializer=nn.initializer.Constant(1.0))
         self.variance_epsilon = eps
 
     def forward(self, hidden_states):
@@ -317,10 +314,8 @@ class T5LayerNorm(nn.Layer):
         # w/o mean and there is no bias. Additionally we want to make sure that the accumulation for
         # half-precision inputs is done in fp32
 
-        variance = paddle.pow(hidden_states.cast(paddle.float32), 2).mean(
-            axis=-1, keepdim=True)
-        hidden_states = hidden_states * paddle.rsqrt(variance +
-                                                     self.variance_epsilon)
+        variance = paddle.pow(hidden_states.cast(paddle.float32), 2).mean(axis=-1, keepdim=True)
+        hidden_states = hidden_states * paddle.rsqrt(variance + self.variance_epsilon)
 
         # convert into half-precision if necessary
         if self.weight.dtype == paddle.float16:
@@ -335,9 +330,9 @@ class NewGELUActivation(nn.Layer):
     """
 
     def forward(self, input: paddle.Tensor) -> paddle.Tensor:
-        return (0.5 * input * (1.0 + paddle.tanh(
-            math.sqrt(2.0 / math.pi) *
-            (input + 0.044715 * paddle.pow(input, 3.0)))))
+        return (
+            0.5 * input * (1.0 + paddle.tanh(math.sqrt(2.0 / math.pi) * (input + 0.044715 * paddle.pow(input, 3.0))))
+        )
 
 
 class T5FiLMLayer(nn.Layer):
@@ -347,8 +342,7 @@ class T5FiLMLayer(nn.Layer):
 
     def __init__(self, in_features, out_features):
         super().__init__()
-        self.scale_bias = nn.Linear(
-            in_features, out_features * 2, bias_attr=False)
+        self.scale_bias = nn.Linear(in_features, out_features * 2, bias_attr=False)
 
     def forward(self, x, conditioning_emb):
         emb = self.scale_bias(conditioning_emb)
